+++ chopper/lib/src/utils.dart
@@ -60,3 +60,9 @@ class _Pair<A, B> {
 
   String toString() => '$first=$second';
 }
+
+bool isTypeOf<ThisType, OfType>() => _Instance<ThisType>() is _Instance<OfType>;
+
+class _Instance<T> {
+  //
+}